Backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.0
python-decouple==3.8
gunicorn==21.2.0
